Prevent user emails being cut off
Make heading break on whitespace - it means if the combination
of name and email address is too long for the line, email address
will go underneath and be more visible.

If email address too big for the allotted space, it will break onto
the next line and all of it will be visible.