Remove issue related to focus shifting onload
The Digital Accessibility Centre (DAC) reported
that it was confusing that focus was moved to the
templates and folders search textbox when the page
loaded on the templates page.

We noted that this issue wasn't reported on the
page where you enter your SMS 2-factor code.

From this, we determined that pages with a single
task could benefit from focus moving to the part
of the page that lets you do that task.

Because of this we:
- found all the pages we shift focus on when they
  load
- kept it for pages where the focus shift is
  important for what users do on the page
- removed it for all other pages (including
  templates)

The work was completed in:

https://github.com/alphagov/notifications-admin/pull/3622

This work also changed the label of the focused
element as a hack to get the page title read out.

This normally happens on load to indicate a
navigation has happened but didn't in Voiceover
when we focused an element on load.

After testing with JAWS and NVDA we found only
Voiceover does this. As Voiceover has the smallest
market share, and navigation can be inferred by
other mean, we removed it in:

https://github.com/alphagov/notifications-admin/pull/3673

That was deployed today.