Remove the styleguide
The idea was that this would be a place to document all the design
patterns used in Notify. However it hasn’t been kept up to date, and,
looking at the `git blame`[1] no new patterns have been added for 5
years.

I think it’s better to get rid of it than have to keep maintaining
something which is inaccurate.

1. https://github.com/alphagov/notifications-admin/blame/64aa0d359c029dddc49f82491756a7a61bd22ded/app/templates/views/styleguide.html