Give the cookie banner an accessible name
Our cookie banner has an accessible description
but no accessible name. It needs to have an
accessible name to distinguish it from other
elements on the page with role=region.

There's some advice in the WCAG guidance that
explains this more:

https://www.w3.org/TR/WCAG20-TECHS/ARIA11.html

This issue was raised in the accessibility review
done as part of our service assessment.

These changes borrow the approach from www.gov.uk of
moving to a label and losing the accessible
description. It also uses the 'cookie banner'
label for consistency with www.gov.uk.